refactor: Move rest of the pages to the right place
